# perf(gatsby): support `elemMatch` as fast filter