--- a/packages/gatsby/src/db/__tests__/fixtures/ensure-loki.js
+++ b/packages/gatsby/src/db/__tests__/fixtures/ensure-loki.js
@@ -4,5 +4,8 @@ module.exports = () => {
   if (backend === `loki`) {
     const lokiDb = require(`../../loki`)
     beforeAll(lokiDb.start)
+    return true
   }
+
+  return false
 }