--- a/packages/gatsby/src/redux/nodes.ts
+++ b/packages/gatsby/src/redux/nodes.ts
@@ -330,7 +330,7 @@ function addNodeToBucketWithElemMatch(
           nestedQuery.path,
           filterCache,
           resolvedNodesCache,
-          valueAtCurrentStep
+          elem
         )
       }
     })