--- a/packages/gatsby/src/schema/node-model.js
+++ b/packages/gatsby/src/schema/node-model.js
@@ -78,14 +78,14 @@ class LocalNodeModel {
    * Replace the cache either with the value passed on (mainly for tests) or
    * an empty new Map.
    *
-   * @param {undefined | Map<string, Map<string, Set<Node>> | Map<string, Node>>} map
+   * @param {undefined | null | FiltersCache} map
    *   (This cached is used in redux/nodes.js and caches a set of buckets (Sets)
    *   of Nodes based on filter and tracks this for each set of types which are
    *   actually queried. If the filter targets `id` directly, only one Node is
-   *   cached instead of a Set of Nodes.
+   *   cached instead of a Set of Nodes. If null, don't create or use a cache.
    */
   replaceTypeKeyValueCache(map = new Map()) {
-    this._typedKeyValueIndexes = map // See redux/nodes.js for usage
+    this._filtersCache = map // See redux/nodes.js for usage
   }
 
   withContext(context) {
@@ -237,7 +237,7 @@ class LocalNodeModel {
       gqlType,
       resolvedFields: fieldsToResolve,
       nodeTypeNames,
-      typedKeyValueIndexes: this._typedKeyValueIndexes,
+      filtersCache: this._filtersCache,
       stats,
     })