The filter for `elemMatch` is basically a sub-match on elements of an array. So rather than matching one explicit path, it matches any of the n paths, splitting off at an array. Other than testing paths through each element in an array, it behaves the same as a regular filter. The `elemMatch` filters can be nested.

These filters weren't before optimized and couldn't take the fast index path. This PR changes that.

## This PR improves `elemMatch` perf at scale
# A baseline 50k site went from 42 q/s to 547 q/s
## (or runtime: from 1200s to 90s)

I've modified "the ifsc" benchmark site and added a silly simple array with three elements to the nodes. Then I modified the query to add an `elemMatch` (aside the already existing `eq` filter on `id`) and ran the before-after benchmark on 500, 5.000, 10.000, and 50.000 pages. The results were as stunning as when first introducing the indexed filters :rocket: See below.

This PR refactors the typing a bit, the first couple of commits are refactorings. Then the meat of the commit is there, with a few fixes in followup commits. It also ports an extended query test suite from a research branch. And it modifies the run query tests to test with and without cache. This way we can at least confirm that the results from using fast indexes are identical to not using the fast indexes at all (which is always a fallback, anyways).

Before:

```
CI=1 NBR=1 N=50000 M=8000 yarn bench
success run queries - 1208.516s - 50002/50002 41.37/s
success Building static HTML for pages - 51.533s - 50002/50002 970.30/s
info Done building in 1290.651 sec

CI=1 NBR=1 N=10000 M=8000 yarn bench
success run queries - 40.606s - 10002/10002 246.32/s
success Building static HTML for pages - 11.668s - 10002/10002 857.21/s
info Done building in 60.489 sec

CI=1 NBR=1 N=5000 M=8000 yarn bench
success run queries - 15.147s - 5002/5002 330.23/s
success Building static HTML for pages - 6.432s - 5002/5002 777.71/s
info Done building in 27.066 sec

CI=1 NBR=1 N=500 M=8000 yarn bench
success run queries - 5.034s - 502/502 99.73/s
success Building static HTML for pages - 1.305s - 502/502 384.67/s
success onPostBuild - 0.000s
info Done building in 8.803 sec

CI=1 NBR=1 N=500 M=8000 yarn bench
never ending
```

After:

```
CI=1 NBR=1 N=50000 M=8000 yarn bench
success run queries - 91.352s - 50002/50002 547.35/s
success Building static HTML for pages - 53.323s - 50002/50002 937.73/s
info Done building in 175.43 sec

CI=1 NBR=1 N=10000 M=8000 yarn bench
success run queries - 11.481s - 10002/10002 871.17/s
success Building static HTML for pages - 11.572s - 10002/10002 864.36/s
info Done building in 31.23 sec

CI=1 NBR=1 N=5000 M=8000 yarn bench
success run queries - 8.036s - 5002/5002 622.45/s
success Building static HTML for pages - 6.409s - 5002/5002 780.42/s
info Done building in 19.8 sec

CI=1 NBR=1 N=500 M=8000 yarn bench
success run queries - 4.959s - 502/502 101.23/s
success Building static HTML for pages - 1.312s - 502/502 382.70/s
info Done building in 8.701 sec

CI=1 NBR=1 N=100000 M=8000 yarn bench
success run queries - 501.655s - 100002/100002 199.34/s
success Building static HTML for pages - 99.309s - 100002/100002 1006.98/s
info Done building in 663.366 sec
```