@freiksenet 2020-04-08
> One particular thing feels missing: test if query actually uses filterWithoutSift or filterWithSift in node-model tests. We just won't notice in tests if the code path switches due to some change in the future. Maybe mock filterWithoutSift or filterWithSift function and expect it was/wasn't called for a particular scenario?

I tried implementing those at some point and gave up once I got too deep into mocking rabbithole. It would be cool to be able to test that, but I'm unsure how atm.