@pvdz 2020-04-08
Published in gatsby@2.20.14

If it regressed, it should be easy to bisect.